requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.11.0
appnope==0.1.4
asgiref==3.8.1
asttokens==3.0.0
attrs==24.3.0
backcall==0.2.0
billiard==4.2.2
build==1.2.2.post1
celery==5.3.6
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cron_descriptor==2.0.6
decorator==5.2.1
distro==1.9.0
Django==4.2.17
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-cors-headers==4.6.0
django-extensions==3.2.3
django-filter==25.1
django-timezone-field==7.1
djangorestframework==3.15.2
easyocr==1.7.2
et_xmlfile==2.0.0
executing==2.2.1
filelock==3.16.1
fsspec==2025.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
imageio==2.37.0
ipython==8.12.3
jedi==0.19.2
jieba==0.42.1
jieba3k==0.35.1
Jinja2==3.1.6
jiter==0.11.0
jupyterlab_pygments==0.3.0
kombu==5.5.4
lazy_loader==0.4
MarkupSafe==3.0.3
matplotlib-inline==0.1.7
mpmath==1.3.0
networkx==3.5
ninja==1.13.0
numpy==2.0.2
openai==1.58.1
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
parso==0.8.5
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.3.0
prompt_toolkit==3.0.52
protobuf==5.29.5
psycopg2==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyclipper==1.3.0.post6
pydantic==2.11.9
pydantic_core==2.33.2
Pygments==2.18.0
pyproject_hooks==1.2.0
python-bidi==0.6.6
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.3
redis==5.0.1
requests==2.32.5
scikit-image==0.25.2
scipy==1.16.2
shapely==2.1.2
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
stack-data==0.6.3
stopwordsiso==0.6.1
sympy==1.14.0
tifffile==2025.9.30
torch==2.8.0
torchvision==0.23.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
